create cl-server and cl-bundle-manager entry_points
