+++ docker/compose_files/docker-compose.yml
@@ -50,7 +50,7 @@ x-codalab-server: &codalab-server
 services:
   rest-server:
     image: codalab/server:${CODALAB_VERSION}
-    command: cl server
+    command: cl-server
     <<: *codalab-base
     <<: *codalab-server
     depends_on:
@@ -58,7 +58,7 @@ services:
 
   bundle-manager:
     image: codalab/server:${CODALAB_VERSION}
-    command: cl bundle-manager
+    command: cl-bundle-manager
     <<: *codalab-base
     <<: *codalab-server
     depends_on:
