+++ setup.py
@@ -34,6 +34,8 @@ setup(
     entry_points={
         'console_scripts': [
             'cl=codalab.bin.cl:main',
+            'cl-server=codalab.bin.server:main',
+            'cl-bundle-manager=codalab.bin.bundle_manager:main',
             'codalab-service=codalab_service:main',
             'cl-worker=codalab.worker.main:main',
             'cl-competitiond=scripts.competitiond:main',
